This PR solves an algorithmic limitation of the kernel by allowing to lookup a rel variable in O(log n) instead of O(n). To this end, we introduce skewed lists in the environment type so as to get O(log n) lookup while keeping a O(1) rel-pushing. In some sense, this is the rel equivalent of PR #263.

We also use the same structure in the environment used by kernel reduction. Before this patch, it was using an array of options instead. In theory array indexing is O(1) and thus faster than the O(log n) one of skewed lists, but in practice the O(n) rebuilding of the array for each call to the reduction was extremely slow and accounting by far of the majority of the time spent in rel-intensive code. As we now share the same environment, it's just a constant-time pointer copy.

Here is a piece of code that displays the algorithmic limitation (courtesy of @JasonGross):
```Coq
Definition Let_In {A P} (v : A) (f : forall x : A, P x) : P v
  := let x := v in f x.

Fixpoint big (a : nat) (sz : nat) : nat
  := match sz with
     | O => a
     | S sz' => Let_In (a * a) (fun a' => big a' sz')
     end.

Ltac do_time n :=
  try (
      once (assert (exists e, e = big 1 n);
            [ lazy [big];
              time pattern Nat.mul, S, O, (@Let_In nat (fun _ => nat))
            | ]);
      fail).

Goal True.
Proof.
do_time 16384.
Abort.
```

Without the patch, this runs about 13s on my machine, while it only takes 0.7s with the patch.

Edit:
Fixes #6502.