@ppedrot 2017-12-26
Here is a benchmark:
```
┌──────────────────────────┬─────────────────────────┬─────────────────────────────────────┬───────────────────────────────────────┬─────────────────────────┬─────────────────┐
│                          │      user time [s]      │             CPU cycles              │           CPU instructions            │  max resident mem [KB]  │   mem faults    │
│                          │                         │                                     │                                       │                         │                 │
│             package_name │     NEW     OLD PDIFF   │           NEW           OLD PDIFF   │            NEW            OLD PDIFF   │     NEW     OLD PDIFF   │ NEW OLD PDIFF   │
├──────────────────────────┼─────────────────────────┼─────────────────────────────────────┼───────────────────────────────────────┼─────────────────────────┼─────────────────┤
│                coq-sf-lf │   15.68   15.86 -1.13 % │   43461839569   43524791719 -0.14 % │    60967932924    60813580147 +0.25 % │  417808  418232 -0.10 % │   0   0  +nan % │
├──────────────────────────┼─────────────────────────┼─────────────────────────────────────┼───────────────────────────────────────┼─────────────────────────┼─────────────────┤
│               coq-sf-plf │   42.30   42.61 -0.73 % │  117827972122  118021041481 -0.16 % │   155022050182   155292399283 -0.17 % │  517752  517808 -0.01 % │   0   0  +nan % │
├──────────────────────────┼─────────────────────────┼─────────────────────────────────────┼───────────────────────────────────────┼─────────────────────────┼─────────────────┤
│    coq-mathcomp-solvable │  191.19  192.37 -0.61 % │  532071506829  534682746986 -0.49 % │   773848188034   778283198562 -0.57 % │  711424  712180 -0.11 % │   0   0  +nan % │
├──────────────────────────┼─────────────────────────┼─────────────────────────────────────┼───────────────────────────────────────┼─────────────────────────┼─────────────────┤
│         coq-fiat-parsers │  655.10  658.56 -0.53 % │ 1828051230601 1838078642148 -0.55 % │  2970298871147  2968653792461 +0.06 % │ 3396948 3523596 -3.59 % │   0   0  +nan % │
├──────────────────────────┼─────────────────────────┼─────────────────────────────────────┼───────────────────────────────────────┼─────────────────────────┼─────────────────┤
│           coq-coquelicot │   75.48   75.79 -0.41 % │  208303267989  208102504218 +0.10 % │   261245076678   261249891717 -0.00 % │  657484  658244 -0.12 % │   0   0  +nan % │
├──────────────────────────┼─────────────────────────┼─────────────────────────────────────┼───────────────────────────────────────┼─────────────────────────┼─────────────────┤
│   coq-mathcomp-odd_order │ 1372.21 1377.08 -0.35 % │ 3820338587737 3837766345041 -0.45 % │  6680023010752  6724965882088 -0.67 % │ 1053024 1111976 -5.30 % │   0   0  +nan % │
├──────────────────────────┼─────────────────────────┼─────────────────────────────────────┼───────────────────────────────────────┼─────────────────────────┼─────────────────┤
│               coq-geocoq │ 3072.69 3082.53 -0.32 % │ 8573471394445 8600962444835 -0.32 % │ 14291365322352 14336892828553 -0.32 % │ 1248456 1252324 -0.31 % │   0   0  +nan % │
├──────────────────────────┼─────────────────────────┼─────────────────────────────────────┼───────────────────────────────────────┼─────────────────────────┼─────────────────┤
│             coq-compcert │  806.84  809.38 -0.31 % │ 2247474859616 2257661056175 -0.45 % │  3416149971979  3420732190589 -0.13 % │ 1375300 1346648 +2.13 % │   0   0  +nan % │
├──────────────────────────┼─────────────────────────┼─────────────────────────────────────┼───────────────────────────────────────┼─────────────────────────┼─────────────────┤
│              coq-unimath │ 1295.32 1299.29 -0.31 % │ 3611809617999 3620151846361 -0.23 % │  6259417805527  6265564405259 -0.10 % │ 1149352 1075068 +6.91 % │   0   0  +nan % │
├──────────────────────────┼─────────────────────────┼─────────────────────────────────────┼───────────────────────────────────────┼─────────────────────────┼─────────────────┤
│     coq-mathcomp-algebra │  171.00  171.47 -0.27 % │  475297188610  475788202281 -0.10 % │   675045651381   676082140834 -0.15 % │  587772  586836 +0.16 % │   0   0  +nan % │
├──────────────────────────┼─────────────────────────┼─────────────────────────────────────┼───────────────────────────────────────┼─────────────────────────┼─────────────────┤
│          coq-fiat-crypto │ 3557.46 3563.84 -0.18 % │ 9903601444971 9923275704184 -0.20 % │ 16723569504223 16755495717230 -0.19 % │ 3198260 3347224 -4.45 % │   0   0  +nan % │
├──────────────────────────┼─────────────────────────┼─────────────────────────────────────┼───────────────────────────────────────┼─────────────────────────┼─────────────────┤
│                coq-color │  562.90  563.46 -0.10 % │ 1576142604655 1576374293817 -0.01 % │  1975732320171  1981689071059 -0.30 % │ 1439552 1439880 -0.02 % │   0   0  +nan % │
├──────────────────────────┼─────────────────────────┼─────────────────────────────────────┼───────────────────────────────────────┼─────────────────────────┼─────────────────┤
│   coq-mathcomp-character │  266.56  266.70 -0.05 % │  740853008567  741335534893 -0.07 % │  1096064114990  1099432190369 -0.31 % │  982720  982760 -0.00 % │   0   0  +nan % │
├──────────────────────────┼─────────────────────────┼─────────────────────────────────────┼───────────────────────────────────────┼─────────────────────────┼─────────────────┤
│ coq-mathcomp-real_closed │  168.10  168.16 -0.04 % │  466507572658  466773175457 -0.06 % │   716226878135   717064153429 -0.12 % │  727756  728496 -0.10 % │   0   0  +nan % │
├──────────────────────────┼─────────────────────────┼─────────────────────────────────────┼───────────────────────────────────────┼─────────────────────────┼─────────────────┤
│    coq-mathcomp-fingroup │   60.23   60.24 -0.02 % │  166251423155  166733024607 -0.29 % │   232855396990   233364426776 -0.22 % │  526152  532420 -1.18 % │   0   0  +nan % │
├──────────────────────────┼─────────────────────────┼─────────────────────────────────────┼───────────────────────────────────────┼─────────────────────────┼─────────────────┤
│   coq-mathcomp-ssreflect │   39.60   39.60 +0.00 % │  109063707338  109425270470 -0.33 % │   135215296138   135452232775 -0.17 % │  477548  477388 +0.03 % │   0   0  +nan % │
├──────────────────────────┼─────────────────────────┼─────────────────────────────────────┼───────────────────────────────────────┼─────────────────────────┼─────────────────┤
│              coq-bignums │   74.40   74.27 +0.18 % │  206872049040  206488901132 +0.19 % │   281122710864   281784886092 -0.23 % │  525404  526908 -0.29 % │   0   0  +nan % │
├──────────────────────────┼─────────────────────────┼─────────────────────────────────────┼───────────────────────────────────────┼─────────────────────────┼─────────────────┤
│       coq-mathcomp-field │  450.91  450.11 +0.18 % │ 1253664114716 1252686778697 +0.08 % │  2098329226276  2101491918038 -0.15 % │  723128  724244 -0.15 % │   0   0  +nan % │
├──────────────────────────┼─────────────────────────┼─────────────────────────────────────┼───────────────────────────────────────┼─────────────────────────┼─────────────────┤
│                coq-flocq │   58.98   58.83 +0.25 % │  161245613618  161923444376 -0.42 % │   207495690937   208803089291 -0.63 % │  645984  644876 +0.17 % │   0   0  +nan % │
├──────────────────────────┼─────────────────────────┼─────────────────────────────────────┼───────────────────────────────────────┼─────────────────────────┼─────────────────┤
│            coq-fiat-core │  100.97  100.69 +0.28 % │  279052486839  279677590274 -0.22 % │   363593129275   363770452526 -0.05 % │  500152  500848 -0.14 % │   0   0  +nan % │
└──────────────────────────┴─────────────────────────┴─────────────────────────────────────┴───────────────────────────────────────┴─────────────────────────┴─────────────────┘


PDIFF = proportional difference between measurements done for the NEW and the OLD Coq version
      = (NEW_measurement - OLD_measurement) / OLD_measurement * 100%

NEW = 1c8516899d60336e971daac74371bd4ba779d8d6
OLD = dea75d74c222c25f6aa6c38506ac7a51b339e9c6
```